End working with quadratic algorithm

+++ linearAlgo.go
@@ -2,6 +2,6 @@ package main
 
 import "fmt"
 
-func linearEquation(b float32, c float32) {
+func linearEquation(b float64, c float64) {
 	fmt.Println(b, c)
 }
